backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
redis==5.0.1
aiohttp==3.9.3
python-dotenv==1.0.1